Fix the mother of all order dependent spec failures
