+++ lib/handcuffs/logger.rb
@@ -4,8 +4,8 @@ module Handcuffs::Logger
     start_time = Time.now
     super
     end_time = Time.now
-    phase = self.class.handcuffs_phase || Handcuffs.config.default_phase
-    log(version, phase, direction, start_time, end_time)
+    local_phase = self.class.handcuffs_phase || Handcuffs.config.default_phase
+    log(version, local_phase, direction, start_time, end_time)
   end
 
   private
@@ -26,6 +26,7 @@ module Handcuffs::Logger
         file.puts(hash.to_json)
       end
     end
+  rescue
   end
 
   def filename
